adjust types.
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@1371 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/switch_rtp.c
@@ -712,7 +712,7 @@ static int rtp_common_read(switch_rtp_t *rtp_session, switch_payload_t *payload_
 		} 
 
 		if (bytes > 0) {
-			int effective_size = bytes - sizeof(srtp_mini_hdr_t);
+			uint32_t effective_size = (uint32_t)(bytes - sizeof(srtp_mini_hdr_t));
 
 			if (rtp_session->recv_msg.header.pt == RTP_MAGIC_NUMBER) {
 				if (!switch_test_flag(rtp_session, SWITCH_RTP_FLAG_MINI)) {
